+++ lib/core/constants/strings.dart
@@ -1,0 +1,5 @@
+class Strings {
+  // App
+  static const appTitle = 'Flutter assignment';
+  static const appVersion = '1.0.1';
+}