+++ CHANGELOG.md
@@ -23,12 +23,29 @@ Here we write the updates of future releases.
 ### Security
 
 
+## [1.0.1] - 15-02-2023
+
+### Added
+- Assets folder to store images
+- Niveus logo in assets folder
+- Core and Presentation main folders
+- Constants, Screens and Utils sub-folders
+- Constant files Bools, Enums, Images and Strings
+- Screens Home screen and Splash screen
+- Utility files Custom text, Custom button and Custom print
+- Google fonts and Sizer plugin in "pubspec.yaml" file
+
+### Changed
+- Custom apk name in "android/app/build.gradle" as "Flutter assignment V1.0.1.apk" which will hereby reflect the latest build version
+- Home widget for material app in "main.dart" file as Splash screen
+
+
 ## [1.0.0] - 15-02-2023
 
 ### Added
 - Changelog file
 - Github repo and connected it
-- Information about the objectives and the link to the prototype in the ReadME.md file
+- Information about the objectives and the link to the prototype in the "README.md" file
 
 ### Changed
 - Changed pubspec file by deleting the commented lines